+++ src/main/Anodyne-Domain/Events/IAggregateEvent.cs
@@ -19,6 +19,6 @@ namespace Kostassoid.Anodyne.Domain.Events
     {
         IAggregateRoot Aggregate { get; }
         object AggregateId { get; }
-        int AggregateVersion { get; }
+        long AggregateVersion { get; }
     }
 }
